fix test hanging
